fix: Update OrdinalPosition type to uint64 in MySqlSchemaProvider.fs

+++ src/SqlHydra.Cli/MySql/MySqlSchemaProvider.fs
@@ -49,7 +49,7 @@ let getSchema (cfg: Config) : Schema =
                 TableName = col["TABLE_NAME"] :?> string
                 ColumnName = col["COLUMN_NAME"] :?> string
                 ProviderTypeName = col["DATA_TYPE"] :?> string
-                OrdinalPosition = col["ORDINAL_POSITION"] :?> int
+                OrdinalPosition = col["ORDINAL_POSITION"] :?> uint64
                 IsNullable =
                     match col["IS_NULLABLE"] :?> string with
                     | "YES" -> true
